requirements.txt
faiss-cpu
numpy
pandas
scikit-learn
sentence-transformers
spacy
pdfminer.six
regex
yake
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
streamlit
transformers
torch
accelerate

